requirements.txt
asgiref==3.9.1
# debugpy==1.8.16
Django==4.2.23
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
gunicorn==23.0.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
faker==18.9.0
